requirements.dev.txt
-r requirements.txt

ruff==0.8.4
pytest==8.3.4
pytest-django==4.9.0
